code from noddy to pigfarm
